cambios

+++ application/controllers/Web.php
@@ -301,32 +301,49 @@ class Web extends CI_Controller {
     $user=$this->input->post('us');//user
     $nuevaContraseña= $this->input->post('contrasena');
     $VerificarContraseña= $this->input->post('vContraseña');
-     $this->db->from('restaurarcontraseña');
-      $this->db->select('*');
-      $this->db->where('token', $token);
-      $restaurar= $this->db->get()->row();
-      if( $restaurar != null )
-      { 
-        if(sha1($restaurar->userId)==$user){
-          if ( $nuevaContraseña ==$VerificarContraseña){
-              $data= array('password'=> $nuevaContraseña);
-              $this->db->where('id', $restaurar->userId);
-              $this->db->update('usuario',$data);
-              $this->db->where('token', $token);
-              $this->db->delete('restaurarcontraseña');
-              header('Location: http://localhost/dimqualityWebshop/login ');
-              
+    if($nuevaContraseña != '' || $VerificarContraseña !=''){
+      $this->db->from('restaurarcontraseña');
+        $this->db->select('*');
+        $this->db->where('token', $token);
+        $restaurar= $this->db->get()->row();
+          if( $restaurar != null && sha1($restaurar->userId)==$user){ 
+              if( strlen($nuevaContraseña)>5 && strlen($VerificarContraseña)>5){
+                  if ( $nuevaContraseña ==$VerificarContraseña){
+                      $data= array('password'=> md5($nuevaContraseña));
+                      $this->db->where('id', $restaurar->userId);
+                      $this->db->update('usuario',$data);
+                      $this->db->where('token', $token);
+                      $this->db->delete('restaurarcontraseña');
+                      $titulo = "Dimquality::WebShop -Recuperar tu contraseña";
+                      $dataHeader['titlePage'] = $titulo;
+                      $dataBody['msg'] ="Su contraseña ha sido cambiada con éxito";
+                      $this->load->view('web/header', $dataHeader);
+                      $this->load->view('web/password_reset', $dataBody); 
+                      
+                  }else{
+                    echo "Hubo un error al procesar su requerimiento.Las contraseñas no coinciden";
+                  }
+              }else{
+                echo "Hubo un error al procesar su requerimiento.La nueva contraseña debe tener minimo 6 caracteres";
+              }
           }else{
-            echo "contraseña no coincide";
+              $this->mensaje();
           }
-        }else{
-           echo "no coincide los usuarios";
-        }
-      }else{
-         echo "no hay";
-      }
+    }else{
+        echo "Hubo un error al procesar su requerimiento.La contraseña no puede ser vacia";
+        
+    }
 
   }
+  
+  public function mensaje(){
+              $titulo = "Dimquality::WebShop -Recuperar tu contraseña";
+              $dataHeader['titlePage'] = $titulo;
+              $dataBody['msg'] ="Hubo un error al procesar su requerimiento";
+              $this->load->view('web/header', $dataHeader);
+              $this->load->view('web/password_reset', $dataBody);
+  }
+  
   //funcion para generar un token para que el usuario pueda cambiar la contraseña
   public function GenerarToken($usuario)
   { 
@@ -339,7 +356,7 @@ class Web extends CI_Controller {
           );
           $resultado=$this->db->insert('restaurarcontraseña', $data);
           if($resultado){
-            $enlace=$_SERVER["SERVER_NAME"].'/dimqualityWebshop/web/ChangePassword?idusuario='.sha1($usuario->id).'&token='.$token;
+            $enlace=$enlace=base_url('/web/ChangePassword?idusuario='.sha1($usuario->id).'&token='.$token);
             return $enlace;
           }else{ 
              return False;
@@ -348,21 +365,34 @@ class Web extends CI_Controller {
 
   public function verificarCorreo(){
       $email = $this->input->post('email');
-      if ($email != ""){
+      $emailRegexResult = preg_match("/^(([A-Za-z0-9]+_+)|([A-Za-z0-9]+\-+)|([A-Za-z0-9]+\.+)|([A-Za-z0-9]+\++))*[A-Za-z0-9]+@((\w+\-+)|(\w+\.))*\w{1,63}\.[a-zA-Z]{2,6}$/", $email);
+      if ($email != "" && $emailRegexResult){
         $this->db->from('usuario');
         $this->db->select('*');
         $this->db->where('email', $email);
         $usuario = $this->db->get()->row();
         if( $usuario != null){
-            $enlace= $this->GenerarToken($usuario);
-            //$this->enviarEmail($usuario->email,$enlace);
-            echo $enlace;
+              $this->db->from('restaurarcontraseña');
+              $this->db->select('*');
+              $this->db->where('userId', $usuario->id);
+              $restaurar= $this->db->get()->row();
+              if( $restaurar == null )
+              { 
+                $enlace= $this->GenerarToken($usuario);
+                //$this->enviarEmail($usuario->email,$enlace);
+              }else{
+                $this->db->where('userId', $usuario->id);
+                $this->db->delete('restaurarcontraseña');
+                $enlace= $this->GenerarToken($usuario);
+              }
+              echo $enlace;
+
         }else{
-              $mensaje= 'este correo no registra ninguna cuenta';
+              $mensaje= 'La dirección de correo proporcionada no está vinculada a ninguna cuenta de usuario';
               echo $mensaje;
         }
       }else {
-          echo "ingrese un correo electronico";
+          echo "Es necesario que ingrese un dirección de correo para recuperar su contraseña";
       }
   }
 
@@ -390,12 +420,5 @@ class Web extends CI_Controller {
           //Se envia el correo al usuario
           mail($email, "Recuperar contraseña", $mensaje, $cabeceras);
     }
-    
-    public function passwordReset(){
-          $titulo = "Dimquality::WebShop -Recuperar tu contraseña";
-          $dataHeader['titlePage'] = $titulo;
-          //$this->load->view('web/password_reset', $dataHeader);
-    }
-  
 }
